[feat] : smartapp.js + server.js 통합
- 두 파일간에 데이터 공유가 자유롭지 못해 하나로 합쳐 개발하는 작업 진행
- imageURL정보를 얻어와서 /api/image URL에 저장함으로써 프론트에서 접근할 수 있도록 구현
